modify dashboard to fix build error regarding useSearchParams() should be wrapped in a suspense boundary at page /dashboard; clear cache and rebuild to new package-lock.jsonto fix styling errors
